utility: move hpc_locks to utility
